+++ core/sail/leveldb/src/main/java/org/eclipse/rdf4j/sail/leveldb/model/NativeBNode.java
@@ -24,13 +24,13 @@ public class NativeBNode extends SimpleBNode implements NativeResource {
 
 	private volatile ValueStoreRevision revision;
 
-	private volatile int internalID;
+	private volatile long internalID;
 
 	/*--------------*
 	 * Constructors *
 	 *--------------*/
 
-	protected NativeBNode(ValueStoreRevision revision, int internalID) {
+	protected NativeBNode(ValueStoreRevision revision, long internalID) {
 		super();
 		setInternalID(internalID, revision);
 	}
@@ -39,7 +39,7 @@ public class NativeBNode extends SimpleBNode implements NativeResource {
 		this(revision, nodeID, UNKNOWN_ID);
 	}
 
-	public NativeBNode(ValueStoreRevision revision, String nodeID, int internalID) {
+	public NativeBNode(ValueStoreRevision revision, String nodeID, long internalID) {
 		super(nodeID);
 		setInternalID(internalID, revision);
 	}
@@ -49,7 +49,7 @@ public class NativeBNode extends SimpleBNode implements NativeResource {
 	 *---------*/
 
 	@Override
-	public void setInternalID(int internalID, ValueStoreRevision revision) {
+	public void setInternalID(long internalID, ValueStoreRevision revision) {
 		this.internalID = internalID;
 		this.revision = revision;
 	}
@@ -60,7 +60,7 @@ public class NativeBNode extends SimpleBNode implements NativeResource {
 	}
 
 	@Override
-	public int getInternalID() {
+	public long getInternalID() {
 		return internalID;
 	}
